Refactor anime table data flow JC-659
