+++ apps/angular/src/app/features/anime-dashboard/components/anime-table/anime-table.component.html
@@ -2,9 +2,9 @@
 @if (animeList.length === 0){
 	<p class="anime-list__empty-message"> No anime to display. </p>
 } @else {
-	<table mat-table [dataSource]="animeList" matSort [matSortActive]="(sortOrder$ | async)?.active ?? ''" [matSortDirection]="(sortOrder$ | async)?.direction ?? ''" (matSortChange)="handleSortChange($event)" [trackBy]="trackByAnimeId">
+	<table mat-table [dataSource]="animeList" matSort [matSortActive]="sortParams.active" [matSortDirection]="sortParams.direction" (matSortChange)="sortChange.emit($event)" [trackBy]="trackByAnimeId">
 		<!-- Image Column -->
-		<ng-container [matColumnDef]="animeEnum.ImageSource">
+		<ng-container [matColumnDef]="animeColumnNames.ImageSource">
 			<th mat-header-cell *matHeaderCellDef> Image </th>
 			<td mat-cell *matCellDef="let element">
 				@if (element.imageSrc){
@@ -15,28 +15,28 @@
 			</td>
 		</ng-container>
 		<!-- English Title Column -->
-		<ng-container [matColumnDef]="animeEnum.EnglishTitle">
-			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeEnum.EnglishTitle"> English Title </th>
+		<ng-container [matColumnDef]="animeColumnNames.EnglishTitle">
+			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeColumnNames.EnglishTitle"> English Title </th>
 			<td mat-cell *matCellDef="let element"> {{element.titleEnglish | empty}} </td>
 		</ng-container>
 		<!-- Japanese Title Column -->
-		<ng-container [matColumnDef]="animeEnum.JapaneseTitle">
+		<ng-container [matColumnDef]="animeColumnNames.JapaneseTitle">
 			<th mat-header-cell *matHeaderCellDef> Japanese Title </th>
 			<td mat-cell *matCellDef="let element"> {{element.japaneseTitle | empty}} </td>
 		</ng-container>
 		<!-- Aired Start Column -->
-		<ng-container [matColumnDef]="animeEnum.AiredStart">
-			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeEnum.AiredStart"> Aired Start </th>
+		<ng-container [matColumnDef]="animeColumnNames.AiredStart">
+			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeColumnNames.AiredStart"> Aired Start </th>
 			<td mat-cell *matCellDef="let element"> {{(element.airedStart | date: 'longDate') | empty}} </td>
 		</ng-container>
 		<!-- Type Column -->
-		<ng-container [matColumnDef]="animeEnum.Type">
+		<ng-container [matColumnDef]="animeColumnNames.Type">
 			<th mat-header-cell *matHeaderCellDef> Type </th>
 			<td mat-cell *matCellDef="let element"> {{element.type}} </td>
 		</ng-container>
 		<!-- Status Column -->
-		<ng-container [matColumnDef]="animeEnum.Status">
-			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeEnum.Status"> Status </th>
+		<ng-container [matColumnDef]="animeColumnNames.Status">
+			<th mat-header-cell *matHeaderCellDef [mat-sort-header]="animeColumnNames.Status"> Status </th>
 			<td mat-cell *matCellDef="let element"> {{element.status}} </td>
 		</ng-container>
