+++ apps/angular/src/app/features/anime-dashboard/components/anime-table/anime-table.component.ts
@@ -1,13 +1,13 @@
-import { ChangeDetectionStrategy, Component, inject, Input, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, inject, Input, Output } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatTableModule } from '@angular/material/table';
 import { EmptyPipe } from '@js-camp/angular/shared/pipes/empty.pipe';
-import { Anime, AnimeEnum } from '@js-camp/core/models/anime';
+import { Anime, AnimeFieldEnum } from '@js-camp/core/models/anime';
 import { MatSortModule, Sort } from '@angular/material/sort';
 import { BasicProgressSpinnerComponent } from '@js-camp/angular/shared/components/basic-progress-spinner/basic-progress-spinner.component';
 import { PageParamsService } from '@js-camp/angular/core/services/page-params.service';
 import { SortOrder } from '@js-camp/core/models/sort-order';
-import { BehaviorSubject, take } from 'rxjs';
+import { BehaviorSubject } from 'rxjs';
 
 /** Table of anime list component. */
 @Component({
@@ -18,33 +18,32 @@ import { BehaviorSubject, take } from 'rxjs';
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	imports: [CommonModule, MatTableModule, EmptyPipe, MatSortModule, BasicProgressSpinnerComponent],
 })
-export class AnimeTableComponent implements OnInit {
+export class AnimeTableComponent {
 	/** Anime list. */
 	@Input()
 	public animeList: readonly Anime[] = [];
 
+	/** Initial sort params. */
+	@Input()
+	public sortParams: Sort = { active: '', direction: '' };
+
+	/** Emits new sort value. */
+	@Output()
+	public sortChange = new EventEmitter<Sort>();
+
 	private readonly pageParamsService = inject(PageParamsService);
 
-	/** Anime enum. */
-	protected readonly animeEnum = AnimeEnum;
+	/** Anime column enum. */
+	protected readonly animeColumnNames = AnimeFieldEnum;
 
 	/** Anime page params. */
 	protected readonly animeParams$ = this.pageParamsService.animeParams$;
 
 	/** Columns names. */
-	protected readonly displayedColumns = Object.values(AnimeEnum);
+	protected readonly displayedColumns = Object.values(AnimeFieldEnum);
 
 	/** Sorting order subject. */
 	protected readonly sortOrder$ = new BehaviorSubject({ active: '', direction: '' } as SortOrder);
-
-	/** @inheritdoc */
-	public ngOnInit(): void {
-		this.animeParams$.pipe(
-			take(1),
-		).subscribe(pageParams => {
-			this.sortOrder$.next(this.parseSortOrder(pageParams.sortOrder));
-		});
-	}
 
 	/**
 	 * TrackBy function using Id of anime.
@@ -54,25 +53,5 @@ export class AnimeTableComponent implements OnInit {
 	 */
 	protected trackByAnimeId(_index: number, item: Anime): Anime['id'] {
 		return item.id;
-	}
-
-	/**
-	 * Handles the sort change.
-	 *
-	 * @param sort New sort value.
-	 */
-	protected handleSortChange(sort: Sort): void {
-		this.pageParamsService.changeSortParams(sort);
-	}
-
-	private parseSortOrder(sortOrder: string): SortOrder {
-		const direction = sortOrder.startsWith('-') ? 'desc' : 'asc';
-
-		const activeField = sortOrder.startsWith('-') ? sortOrder.slice(1) : sortOrder;
-
-		return {
-			active: activeField,
-			direction,
-		};
 	}
 }
